raspi_to_linphone/requirements.txt
baresipy
python-dotenv
pip
certifi
pexpect
ptyprocess
urllib3
opentone
pydub
idna

picamera2
av
numpy
pillow
attrs
charset-normalizer
piexif
typing_extensions
jsonschema
libarchive-c
tqdm
videodev2
simplejpeg
pidng
referencing
jsonschema-specifications
typing_extensions